simplify implimentation

+++ collection/priority_queue.go
@@ -9,7 +9,7 @@ import (
 
 type WithPriority[T any] struct {
 	Item     T
-	Priority float32
+	Priority float64
 }
 
 type priorityQueue[T any] []*WithPriority[T]
@@ -48,7 +48,7 @@ func NewPriorityQueue[T any](capacity int) *PriorityQueue[T] {
 	}
 }
 
-func (pq *PriorityQueue[T]) Push(item T, priority float32) {
+func (pq *PriorityQueue[T]) Push(item T, priority float64) {
 	heap.Push(
 		&pq.priorityQueue,
 		&WithPriority[T]{
